requirements.txt
numpy
torch
torchvision
Pillow
moviepy
librosa
transformers
opencv-python-headless
requests
matplotlib
scipy
pyspellchecker
